setting up forms, still routing errors
