+++ app/src/features/surveys/invite/InviteSurveyMembersPage.tsx
@@ -1,0 +1,76 @@
+import LoadingButton from '@mui/lab/LoadingButton';
+import Button from '@mui/material/Button';
+import Container from '@mui/material/Container';
+import Stack from '@mui/material/Stack';
+import PageHeader from 'components/layout/PageHeader';
+import { Formik, FormikProps } from 'formik';
+import { useBiohubApi } from 'hooks/useBioHubApi';
+import { useRef, useState } from 'react';
+import { useHistory } from 'react-router';
+import yup from 'utils/YupSchema';
+import InviteSurveyMembersForm, { IManageUsersFormValues } from './form/InviteSurveyMembersForm';
+
+const initialValues: IManageUsersFormValues = {
+  selectedSurveys: [],
+  selectedMembers: []
+};
+
+const manageUsersYupSchema = yup.object().shape({
+  selectedSurveys: yup.array(yup.number()).min(1, 'You must select at least one survey.'),
+  selectedMembers: yup
+    .array(yup.object().shape({ system_user_id: yup.number(), collection_role_name: yup.string() }))
+    .min(1, 'Invite a team member and assign them a role.')
+});
+
+/**
+ * Page for bulk adding emmbers to surveys
+ *
+ * @return {*}
+ */
+export const InviteSurveyMembersPage: React.FC = () => {
+  const biohubApi = useBiohubApi();
+  const history = useHistory();
+  const formikRef = useRef<FormikProps<IManageUsersFormValues>>(null);
+  const [isSaving] = useState(false);
+  const handleSubmit = async (values: IManageUsersFormValues) => {
+    await biohubApi.survey.addBulkSurveysMembers(values);
+  };
+  //
+
+  return (
+    <>
+      <PageHeader title="Manage Users" />
+      <Container maxWidth="xl" sx={{ py: 3 }}>
+        <Formik
+          innerRef={formikRef}
+          initialValues={initialValues}
+          validateOnBlur={false}
+          validateOnChange={false}
+          validationSchema={manageUsersYupSchema}
+          enableReinitialize={true}
+          onSubmit={handleSubmit}>
+          <InviteSurveyMembersForm />
+
+          <Stack mt={4} flexDirection="row" justifyContent="flex-end" gap={1}>
+            <LoadingButton
+              loading={isSaving}
+              type="submit"
+              color="primary"
+              variant="contained"
+              onClick={() => formikRef.current?.submitForm()}
+              data-testid="submit-manage-users-button">
+              Save and Exit
+            </LoadingButton>
+            <Button
+              disabled={isSaving}
+              color="primary"
+              variant="outlined"
+              onClick={() => history.push('/admin/summary')}>
+              Cancel
+            </Button>
+          </Stack>
+        </Formik>
+      </Container>
+    </>
+  );
+};
