+++ app/src/interfaces/useSurveyApi.interface.ts
@@ -93,7 +93,7 @@ export interface ISurveyMembersAdvancedFilters {
   system_user_id?: number;
   keyword?: string;
 }
-export interface IPostSurveyMember {
+export interface IPostSurveyMember extends ISystemUser {
   system_user_id: number;
   survey_role_name: SURVEY_ROLE;
 }
